fix: changelog and company model

+++ routes/company.js
@@ -14,8 +14,8 @@ router.get("/my-company", veryToken, (req, res) => {
     .catch((err) => res.status(400).json(err));
 });
 
-// Post a company if you are authorized
-router.post("/company", veryToken, (req, res) => {
+// Create your company if you have a user
+router.post("/create-company", veryToken, (req, res) => {
   const company = { ...req.body};
   Company.create(company)
     .then((result) => {
@@ -25,39 +25,30 @@ router.post("/company", veryToken, (req, res) => {
 });
 
 // Delete your company
-router.delete("/delete-company", veryToken, (req, res) => {
-  const { id } = req.body;
-
-  Company.findById(id).then((company) => {
-    if (company.user.equals(req.user._id)) {
-      Company.findOneAndDelete(id)
+router.get("/delete-company/:id", veryToken, (req, res) => {
+  const { _id: userId } = req.user
+  const { id } = req.params;
+      Company.findOneAndDelete({ _id: id, user: userId })
         .then((company) => {
           res.status(200).json({
             result: company,
           });
         }).catch((err) => res.status(400).json(err));
-    } else {
-      res.status(401).json({ error: 'You are not allowed to remove this company' });
-    }
-  }).catch(() => res.status(400).json({ error: 'Company not found!' }));
 });
 
 // Modify your company
-router.patch("/modify-company", veryToken, (req, res) => {
-  const { id } = req.body;
-
-  Company.findById(id).then((company) => {
-    if (company.user.equals(req.user._id)) {
-      Company.findOneAndUpdate(id)
+router.patch("/modify-company/:id", veryToken, (req, res) => {
+  const { _id: userId } = req.user
+  const { id } = req.params;
+      Company.findOneAndUpdate({ _id: id, user: userId })
       .then((company) => {
         res.status(200).json({
           result: company,
         });
     }).catch((err) => res.status(400).json(err));
-  } else {
-    res.status(401).json({ error: 'You are not allowed to modify this company' });
-  }
-}).catch(() => res.status(400).json({ error: 'Company not found!' }));
 });
 
 module.exports = router;
+
+
+
